Adding structure for tileimprovement score

+++ C7Engine/AI/StrategicAI/ExpansionPriority.cs
@@ -60,7 +60,7 @@ namespace C7GameData.AIData {
 			return "ExpansionPriority";
 		}
 
-		private double ApplyEarlyGameMultiplier(int score)
+		private double ApplyEarlyGameMultiplier(double score)
 		{
 			//If it's early game, multiply this score.
 			//TODO: We haven't implemented the part for "how many turns does the game have?" yet.  So this is hard-coded.
@@ -72,7 +72,7 @@ namespace C7GameData.AIData {
 			return score;
 		}
 
-		private double ApplyNationTraitMultiplier(int score, Player player) {
+		private double ApplyNationTraitMultiplier(double score, Player player) {
 			// TODO: The "Expansionist" trait should give a higher priority to this strategic priority.
 			return score;
 		}
